Added an improvement to the Omniglot class
========================================================================

Previously, caching of indices was being done prior to initiating the
class, which was time consuming.  The modification is to add to the cache
as indices are called, so that the computation is spread out.  This is
fine with this class as it's farming out things to the cpus as things
are happening on the gpus.