Model files to .gitignore and a bug fix
========================================================================

Added .pt and .h5 file extensions to .gitignore, so it's not tracking
all the files created during training (or picking them up, which makes
it hard to see whats going on with the status).  Also, fixed a but in
models/matching_networks.py that seemed to be keeping the model from
training well (We'll see about this).  Was using torch CrossEntropyLoss
which is expecting scores, not softmax's.  So this would have been
problematic.

========================================================================

https://pytorch.org/docs/master/nn.html#torch.nn.CrossEntropyLoss
https://stackoverflow.com/questions/49390842/cross-entropy-in-pytorch